Create README.md
High-Content Scan2 (HCScan2) is an ImageJ macro that performs an automated high-content analysis of two-channel fluorescence images. In the interactive mode the user can use the ROI Manager to view the cell/ROI corresponding to each measurement, a feature that facilitates the optimization of assay parameters. An Excel-compatible report and zipped folder containing images of each cell/ROI are generated, allowing users to archive and publish results at the cellular level.